Revert "a TOMAR POR CULO EL AXEL"
This reverts commit 2559f5d2d10226fed3ae5fd97b50e8c570e837e1.

+++ Assets/Scripts/Isra/GameManager.cs
@@ -45,7 +45,7 @@ public class GameManager : MonoBehaviour
         numberOfEnemiesSaved = 0;
     }
 
-    public void SetupLevel(int levelNumber, float levelSeconds, float numberOfTotalEnemies)
+    public void SetupLevel(int levelNumber, float levelSeconds, int numberOfTotalEnemies)
     {
         // Restart the variables
         Restart();
